Add OfficeLayout and apply to office pages
