+++ pages/office/clients/[id].js
@@ -2,7 +2,7 @@
 import React, { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
-import { Layout } from '../../../components/Layout';
+import OfficeLayout from '../../../components/OfficeLayout';
 import { fetchVehicles } from '../../../lib/vehicles';
 
 const EditClientPage = () => {
@@ -60,10 +60,10 @@ const EditClientPage = () => {
     fetchVehicles(id).then(setVehicles);
   };
 
-  if (loading) return <Layout><p>Loading…</p></Layout>;
+  if (loading) return <OfficeLayout><p>Loading…</p></OfficeLayout>;
 
   return (
-    <Layout>
+    <OfficeLayout>
       <h1 className="text-2xl font-semibold mb-4">Edit Client</h1>
       {error && <p className="text-red-500">{error}</p>}
       <form onSubmit={submit} className="space-y-4 max-w-md">
@@ -143,7 +143,7 @@ const EditClientPage = () => {
           </table>
         )}
       </div>
-    </Layout>
+    </OfficeLayout>
   );
 };
