+++ pages/office/job-management/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Layout } from '../../../components/Layout';
+import OfficeLayout from '../../../components/OfficeLayout';
 
 const JobManagementPage = () => (
-  <Layout>
+  <OfficeLayout>
     <h1 className="text-xl font-semibold">Job Management</h1>
     {/* TODO: Kanban or board view */}
-  </Layout>
+  </OfficeLayout>
 );
 
 export default JobManagementPage;
